Naming hack, libssh.so == ssh? Really Android?
Ok, so the NDK build system is an idiot. It literally does not
distinguish between a library named libssh, and an executable named ssh.
So to hackaround this problem, we call the executable as "_ssh" and then
rename it back to "ssh" during the move to the assets/bin path.