Performance improvement
Get slow operations off the main thread, specifically, the function of
the button that checks the state of ssh installation on the system --
start a new thread and run it there, then let it call back to the UI
thread when it has completed.

Also add a scrollview to System fragment.