Change installation target to /data/bin/
The way Android is moving towards a fully locked out system partition
(for security using dm-verity, not lockout), it makes it appropriate to
move installation to /data partition.

This required changing the linking to custom built libraries to all
static, since we no longer have access to a path known to the linker.